Demonstration of pollInterval bug
Run this against GitHunt-API.  Open the network tab. You should see a graphql
request every 5 seconds.

Click the word "short".  That should switch to every half a second.  You can
confirm that the pollInterval prop on the Apollo() component has changed to 500.

Click the word "long".  The pollInterval prop changes back to 5000, but in the
network tab it's still polling every half a second.